test(api): add account-recovery-demand.serializer.test.js
